move tests around to make them findable
